requirements.txt
faster-whisper>=1.0.3
sounddevice>=0.4.6
numpy>=1.24
PySide6>=6.6
anthropic>=0.39
openai>=1.54
google-generativeai>=0.8
alibabacloud-nls-python-sdk>=1.0.0
